Journal and Dialog Integration
After each dialog if any notes or objectives were contained in the dialogData they get added to the Journal in the appropriate sections.
Furthermore, If the player has any stories they could choose to write in the journal, they get displayed as well and theplayer can choose which one to save in the journal.